pyproject.toml
[tool.poetry]
authors = ["Gabriel Manhães"]
name = "library"
version = "1.0.0"
description = ""
readme = "README.md"
packages = [
    { include = "librarian", from = "src" },
    { include = "archiver", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.10"
alembic = "^1.15.2"
requests = "^2.32.3"
celery = "^5.3.0"
redis = "^5.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
httpx = "^0.27.0"

[tool.poetry.scripts]
librarian = "librarian.app:main"
archiver = "archiver.app:main"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[tool.poetry.plugins."poetry.application.plugin"]
export = "poetry_plugin_export.plugin:ExportPlugin"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
